[JABERA-13] Move user repository to auth package
