+++ src/main/java/com/the/dark/side/crew/fejsbuk/mapper/CommentMapper.java
@@ -5,7 +5,7 @@ import com.the.dark.side.crew.fejsbuk.model.PostEntity;
 import com.the.dark.side.crew.fejsbuk.auth.domain.entity.UserEntity;
 import com.the.dark.side.crew.fejsbuk.model.dto.CommentDto;
 import com.the.dark.side.crew.fejsbuk.repository.PostRepository;
-import com.the.dark.side.crew.fejsbuk.repository.UserRepository;
+import com.the.dark.side.crew.fejsbuk.auth.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
