+++ src/main/java/com/the/dark/side/crew/fejsbuk/service/impl/UserServiceImpl.java
@@ -1,7 +1,7 @@
 package com.the.dark.side.crew.fejsbuk.service.impl;
 
 import com.the.dark.side.crew.fejsbuk.auth.domain.entity.UserEntity;
-import com.the.dark.side.crew.fejsbuk.repository.UserRepository;
+import com.the.dark.side.crew.fejsbuk.auth.repository.UserRepository;
 import com.the.dark.side.crew.fejsbuk.service.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
